+++ src/components/common/BasicLogo.vue
